move Icons components out of Geral subfolder
also deletes icons no longer used after UI update
